add registration logic

+++ learning/views.py
@@ -1,3 +1,4 @@
+from django.http import HttpResponse
 from django.shortcuts import redirect, render
 from django.db.models import Q
 from learning.forms import GroupForm
@@ -6,12 +7,17 @@ from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 
 # Create your views here.
 
 def loginPage(request):
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -28,12 +34,29 @@ def loginPage(request):
         else:
             messages.error(request, 'Username or password is Invalid')
 
-    context = { }
+    context = { 'page':page}
     return render(request, 'learning/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerUser(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occurred during registration")
+
+    context = { 'form': form}
+    return render(request, 'learning/login_register.html', context)
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -73,6 +96,9 @@ def updateGroup(request, pk):
     group = Group.objects.get(id = pk)
     form = GroupForm(instance=group)
 
+    if request.user != group.leader:
+        return HttpResponse("You are not authorized")
+
     if request.method == 'POST':
         form = GroupForm(request.POST, instance=group)
         if form.is_valid():
@@ -85,6 +111,9 @@ def updateGroup(request, pk):
 @login_required(login_url='login-page')
 def deleteGroup(request, pk):
     group = Group.objects.get(id=pk)
+
+    if request.user != group.leader:
+        return HttpResponse("You are not authorized")
 
     if request.method == 'POST':
         group.delete()
